Moved fopen_caseless() functions to own file
